pyproject.toml
[project]
name = "amas-code"
version = "0.1.0"
description = "AI coding agent — any model, any hardware, ~760 lines"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "litellm>=1.40",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "beautifulsoup4>=4.12",
    "httpx>=0.27",
]

[project.optional-dependencies]
browser = ["playwright>=1.40"]
parsing = ["tree-sitter>=0.21", "tree-sitter-languages>=1.10"]
all = ["amas-code[browser,parsing]"]

[project.scripts]
amas = "amas_code.amas:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
